+++ lib/oli_web/live/users/authors_view.ex
@@ -145,7 +145,8 @@ defmodule OliWeb.Users.AuthorsView do
              },
              changes
            )
-         )
+         ),
+       replace: true
      )}
   end