+++ lib/oli_web/live/users/users_view.ex
@@ -143,7 +143,8 @@ defmodule OliWeb.Users.UsersView do
              },
              changes
            )
-         )
+         ),
+       replace: true
      )}
   end